rename tests to represent complexity increases
